vel_discrete_points for adaptive grids
The velocity calculated from discrete points is now calculated even for
adaptive grids. The calculation is such different from the method used
for the regular grid. A cube is created (in the same octant like in the
regular grid case) based on the current propGrid cell size (which is
equal) and the velocity points in the corners are calculated based on
the model grid velocity the propGrid cell belongs to. Then the classical
trilinear interpolation is done.

* oct_virtcube: a new sbr to create the virtual cube
* find_dyn_cell1: found a bug in the condition if the packet is out of
  computational domain
* vel_discrete_points: two forks: one for regular and one for adaptive
  grid