debugging line interaction
Still in the debug mode. Into r_kappa_line added a control part of a
calculation of the optical depth in lines using derivative dr/dv. It
should be as the same as the equation possesing R_inf/V_inf. The values
are the same, however there is a large scatter of the calculated values.
Hence the additional testing must take place.

Furthermore I implemented more save dummy packet. Into module
dummypacket were added several subroutines to activate, copy and
deactivate dummypackets. If a dummypacket is already used, it cannot be
used by another sbr, until the current dummypacket is deactivated.

* cmf_freq: deleted input parameter cur_pos
* doppler_factor2: deleted, it is not used anymore
* dummypacket: added several subroutines to dealt with dummypackets
* velo: deleted one parameter