seeking for the n closest points
Still continue in the developing of seeking of n closest points. For
this purpose I added some new sbrs: seek_nclosest_points and a module virt_gridAB
which creates virtual grids A and B and also creates arrays of modGrid
points according to virtual grid index plus another array contains
positions of these modGrid cells in the arrays. It is compiled with no
mistakes, hovewer, it has not been run yet!!

* n_closest_points_2D: a special sbr for 2D finding the N_points closest in the modGrid
* propmodgrid_diagnostics: added calling for a calculation of the
  closest points
* seek_nclosest_points: a new sbr for seeking (and creating) an array
  with n closest points
* virt_gridAB: a new module, it also creates arrays as written above
* virtual_points: some fixes in one letter variables