new emission scheme in do_ipackage
New direction was set up via dummypacket. The dummypacket machinery has
been changed, therefore the calculation must have been changed as well.
The dummypacket is created only in the sbr i_radtrans, after generating
a new direction, the dummypacket is saved and the direction is saved
into a variable new_dir which is sent to the sbr do_ipackage. If
radiative deexcitation occurs, it is used as a new packet direction.

* do_ipackage: added description, added new dummypacket machinery and some
   cleaning work has been done
* emit_rpackage: added dummypacket machinery
* i_radtrans: added dummypacket machinery to generate new random
   direction